extract overline, underline, and rule into their own files
